Add map validation
After parsing, the following verifications are made:
- Empty map;
- No player spawn points;
- Playable area not sealed with walls;
