+++ include/cub3d/cub3d.h
@@ -1,11 +1,9 @@
 #ifndef CUB3D_H
 # define CUB3D_H
 
-# define READ_SIZE	1024
-# define ERROR_MAX	1024
-# define MAP_MAX	1024
-# define ENTITY_MAX	64
+# include <cub3d/types.h>
 
+int	init_scene(t_scene *scene, const char *file);
 int	print_error(const char *type, const char *msg);
 
 #endif
