+++ include/cub3d/strings.h
@@ -7,9 +7,12 @@
 # define SECT_PARSE	"parsing"
 
 # define ERR_EMPTY_LINE	"empty map line"
+# define ERR_MAP_EMPTY	"empty map"
 # define ERR_MAP_MAX	"map too large"
 # define ERR_ENTITY_MAX	"too many entities"
+# define ERR_NO_SPAWN	"no player spawn point"
 # define ERR_MUL_SPAWN	"multiple player spawn points"
 # define ERR_INV_CELL	"invalid map character"
+# define ERR_MAP_SEAL	"playable area not sealed"
 
 #endif
