+++ include/cub3d/types.h
@@ -2,40 +2,8 @@
 # define TYPES_H
 
 # include <mlx_int.h>
-# include <cub3d/cub3d.h>
+# include <cub3d/defines.h>
 # include <cub3d/vector.h>
-
-enum e_cell
-{
-	SPACE,
-	WALL,
-	ENTITY,
-	PLAYER
-};
-
-enum e_entity
-{
-	KEY,
-	DOOR,
-};
-
-enum e_texture
-{
-	NORTH_WALL_TEX,
-	EAST_WALL_TEX,
-	SOUTH_WALL_TEX,
-	WEST_WALL_TEX,
-	KEY_TEX,
-	DOOR_TEX,
-	TEXTURE_COUNT,
-};
-
-enum e_color
-{
-	FLOOR_COL,
-	CEILING_COL,
-	COLOR_COUNT,
-};
 
 typedef union s_color
 {
