+++ src/error.c
@@ -1,23 +1,9 @@
-#include <unistd.h>
-#include <string.h>
 #include <errno.h>
 #include <libft.h>
 #include <cub3d/cub3d.h>
 #include <cub3d/strings.h>
 
-static void	buffer_error(const char *const str)
-{
-	static char	buffer[ERROR_MAX];
-
-	if (str)
-		ft_strlcat(buffer, str, ERROR_MAX);
-	else
-	{
-		ft_strlcat(buffer, "\n", ERROR_MAX);
-		write(2, buffer, ft_strlen(buffer));
-		buffer[0] = '\0';
-	}
-}
+static void	buffer_error(const char *str);
 
 int	print_error(const char *const type, const char *const msg)
 {
@@ -31,4 +17,18 @@ int	print_error(const char *const type, const char *const msg)
 		buffer_error(strerror(errno));
 	buffer_error(NULL);
 	return (1);
+}
+
+static void	buffer_error(const char *const str)
+{
+	static char	buffer[ERROR_MAX];
+
+	if (str)
+		ft_strlcat(buffer, str, ERROR_MAX);
+	else
+	{
+		ft_strlcat(buffer, "\n", ERROR_MAX);
+		write(2, buffer, ft_strlen(buffer));
+		buffer[0] = '\0';
+	}
 }
