+++ src/init.c
@@ -1,4 +1,3 @@
-#include <fcntl.h>
 #include <libft.h>
 #include <cub3d/cub3d.h>
 #include <cub3d/strings.h>
@@ -10,6 +9,10 @@ int	init_scene(t_scene *const scene, const char *const file)
 
 	if (fd == -1)
 		return (print_error(SECT_INIT, NULL));
-	ft_bzero(scene, sizeof(t_scene));
-	return (parse_map(scene, fd));
+	ft_memset(scene->map, WALL, sizeof(char) * MAP_MAX * MAP_MAX);
+	scene->player.pos.x = -1.0f;
+	if (parse_map(scene, fd) || validate_map(scene))
+		return (1);
+	close(fd);
+	return (0);
 }
